UniversalPackagesV0 In-line validation
