+++ Tasks/UniversalPackagesV0/package.json
@@ -23,7 +23,7 @@
     "@types/node": "^24.10.0",
     "@types/uuid": "^8.3.0",
     "azure-pipelines-task-lib": "^5.2.4",
-    "azure-pipelines-tasks-packaging-common": "^3.270.0",
+    "azure-pipelines-tasks-packaging-common": "^3.273.1",
     "azure-pipelines-tasks-utility-common": "^3.270.0"
   },
   "devDependencies": {
