+++ Tasks/UniversalPackagesV0/task.loc.json
@@ -9,7 +9,7 @@
   "category": "Package",
   "version": {
     "Major": 0,
-    "Minor": 272,
+    "Minor": 274,
     "Patch": 0
   },
   "runsOn": [
